added semantic label tag for input tags
