--- public/food_and_bev2.js
@@ -1,12 +1,0 @@
-import {
-    getItemsByCategory,
-    submitHandler,
-    assignActiveClass,
-} from './main.js';
-
-const ITEM_CATEGORY = 'foodAndBev';
-
-assignActiveClass();
-const form = document.querySelector('form');
-form.addEventListener('submit', (e) => submitHandler(e, ITEM_CATEGORY));
-getItemsByCategory(ITEM_CATEGORY);
