--- public/must_tries.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Must-Tries</title>
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-<link href="https://fonts.googleapis.com/css2?family=Lato:wght@100;300;400;700&family=Merriweather:ital@0;1&family=Noto+Serif+Georgian:wght@200;300;500&family=Space+Mono&family=Work+Sans:wght@800&display=swap" rel="stylesheet">
-    <link href="styles.css" rel="stylesheet">
-</head>
-<nav>
-    <ul>
-        <li class="crumb"><a class="crumb-link" href="./">Home</a></li>
-        <li class="crumb"><a class="crumb-link" href="./activities.html">Activities</a></li>
-        <li class="crumb"><a class="crumb-link" href="./food_and_bev.html">Food & Bev</a></li>
-        <li class="crumb"><a class="crumb-link" href="./must_tries.html">Must-Tries</a></li>
-    </ul>
-</nav>
-<body>
-    <header>
-        <h1>Must-Tries</h1>
-        <h2>Add an Aussie must-try</h2>
-    </header>
-<main>
-    <form>
-        <input
-            type="text"
-            id="name"
-            placeholder="Must-try">
-        <input
-            type="text"
-            id="recommended-by"
-            placeholder="Recommended by">
-        <input
-            type="text"
-            id="image-URL"
-            placeholder="Image URL">
-        <button id="add-item-btn"><span>Add</span></button>
-    </form>
-    <section type="text" id="items-container"></section>
-</main>
-    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-    <script type="module" src="main.js"></script>
-    <script type="module"  src="must-tries.js"></script>
-</body>
-</html>
