+++ public/styles.css
@@ -310,3 +310,14 @@ input[type='text'] {
     margin-bottom: 2.5px;
     margin-right: 1px;
 }
+
+.visually-hidden {
+    position: absolute;
+    width: 1px;
+    height: 1px;
+    margin: -1px;
+    padding: 0;
+    overflow: hidden;
+    clip: rect(0, 0, 0, 0);
+    border: 0;
+}
