+++ views/pages/activities.ejs
@@ -26,14 +26,17 @@
 <main>
     <form>
         <div class="inputsContainer">
+            <label for="activity-name" class="visually-hidden"> Enter an Activity name</label>
             <input
                 type="text"
-                id="name"
+                id="activity-name"
                 placeholder="Activity Name">
+            <label for="recommended-by" class="visually-hidden">Recommended by</label>
             <input
                 type="text"
                 id="recommended-by"
                 placeholder="Recommended by">
+            <label for="image-URL" class="visually-hidden">Paste an image URL</label>
             <input
                 type="text"
                 id="image-URL"
