+++ views/pages/must_tries.ejs
@@ -25,14 +25,17 @@
     </header>
 <main>
     <form>
+        <label for="must-tries-name" class="visually-hidden">Enter a must-try title</label>
         <input
             type="text"
-            id="name"
+            id="must-tries-name"
             placeholder="Must-try">
+        <label for="recommended-by" class="visually-hidden">Recommended by</label>
         <input
             type="text"
             id="recommended-by"
             placeholder="Recommended by">
+        <label for="image-URL" class="visually-hidden">Paste an image URL</label>
         <input
             type="text"
             id="image-URL"
